Fixed test; fixed warning
git-svn-id: https://open-ms.svn.sourceforge.net/svnroot/open-ms/OpenMS@3385 6adb6e08-d915-0410-941f-83917bcadc18

+++ include/OpenMS/FORMAT/Base64.h
@@ -294,7 +294,7 @@ namespace OpenMS
 		}
 
 		//reserve enough space in the output vector
-		out.reserve(std::ceil((4.0*src_size)/3.0)+6.0);
+		out.reserve((UInt)(std::ceil((4.0*src_size)/3.0)+6.0));
 
 		// sort all read bytes correctly into a char[4] (double) or
 		// char[8] (Real) and push_back when necessary.
